Task1/requirements.txt
psycopg2-binary
python-dotenv==1.1.1
pytest
python-on-whales